Fix for issue 222.
Revised KeyedResourcePool.close(K key) to fix leaking sockets. Instead of removing the pool itself, we close it. Closing a pool clears all idle resources in the pool at the time of closure. Additionally, the SocketPool sets a timestamp for the SocketDestination object that notes when it was last closed. We also set a timestamp for each Socket that notes when it was created. The validate method of SocketResourceFactory now checks these two timestamps and deems invalid any sockets created prior to the pool closing. Only those sockets created before the closure will be auto-closed on returning to the pool; new sockets will not be closed.