gradle build support for voldemort
1) 'gradle jar' will build binaries that are equivalent to 'ant release'
2) only 'ant release' task is ported to gradle. use 'ant junit' to run
    the unit tests.
3) zip and tar will not be built, unless explicitly requested. by the
   command 'gradle zip' or 'gradle tar' or 'gradle zip tar'

These are the fine level details of the gradle jar
1) The class files has same binary content except for the scala class
files, since it is not used in production, I am ignoring scala as it is
not used in production.

2) The following two lines went missing from the MANIFEST.MF in both
contrib.jar and voldemort.jar
Ant-Version: Apache Ant 1.7.1
Created-By: 20.2-b06 (Sun Microsystems Inc.)

3) some test class files has different binary content but binary
equivalence is not important for them as we dont ship them. Test jar has
some new class files most likely from junit ignoring this one as well.

4) src.jar now contains package.html included as well ( not intentional,
but not problem either so ignoring it)

5) tar and zip tasks are not part of the build, but can be requested
explicility. The contents of tar and zip are going to be little different
than previous but serves the purpose so we can live with it.