DataCleanup job should not run during rebalance
1) There is a check to not start Data Cleanup job during rebalance,
but after it is started, there is no periodic check.  This causes the
data cleanup to compete with rebalance.
2) Check every 10,000 entries scanned to make sure that there is no
rebalance in progress.
3) Added error condition around the data retention to make sure that if
it is set to 0 or lesser days the value is ignored and an error is
logged. Previously if it is set to 0, it will wipe out your entire data.