ReplaceNodeCLI command
The checkin adds a command to replace node. It takes in the old cluster
URL and node id to be replaced and the new cluster URL. It modifies the
old cluster to remove the old node and adds the new node in place. It
automatically restores the data from the other nodes in the old cluster
to the new node.