Share the constant for cluster.xml, stores.xml
Share the same Constant from SystemStoreConstants for stores.xml
cluster.xml and metadata-versions. Previously each class declared its
own constant and searching for the usage within the source was a
nightmare. I had to do multiple file searches to piece together
information. This will help in the future to make the dependencies
easier to track.