Version clock increments on each restart
Voldemort verion clock increments on each restart and hence rejects future
updates. Increment the clock only when necessary. Ideally if the version
value already exists, the function should return early, but not sure when
cluster.xml, stores.xml and for each stores a key was introduced. And what
happens if the keys were not there. So found a safe middle ground to increment
the version when the key is not there.