Read after truncate crashes the JVM
If you truncate a read-only store and send read operations get/getall
for one of the existent keys it crashes the JVM with SegFault. Some of
the data is cached in java collections and it tries to access the memory
mapped file which has been closed by the truncate. This causes the JVM
to crash. Now the java collections are cleared and the code tries to
handle this gracefully.  But this is not an ideal code more of a
workaround.

Fixed spewing of log messages to the standard output, as they are
already captured in the log4j loggers.

Added unit test for the read after truncate scenario.