Nio client thread names for multiple factories
If a Voldemort client creates multiple factories ( either for isolation
or for talking to multiple clusters) the thread names are difficult to
track in logs and stack dumps. Now the thread names will contain the
factory identifiers in the names.

Servers does not pass in Stats object, so the servers hosting this
client will continue to be the same.