Print the summary of the stores.xml change
When you use the new vadmin.sh meta set stores.xml it will print a
summary of all the changes you made to the stores.

It will categorize them to the following changes
1) Warn if New stores are being added
2) Warn if Stores are being removed
3) Unchnaged stores

For the modified stores, it will add them to atleast one category
1) Replication factor changed
2) Schema changed
3) All other changes

This will help you identify the changes to the stores.xml before you
update it.

The vadmin.sh meta check stores.xml will do a sample key fetch against
the store and report if there is an issue.