RouteToAllStrategy routes to node 0..n
RouteToAllStrategy tries the node always in a fixed order.
This creates too much metadata queries on the node 0.
For zoned cluster, the node with lowest id gets bombarded with too many
connections and get queries.

Create a shuffled node, when the cluster is initialized and use this in
the routing strategy. The random seed is used at the initialization to
make it random every time the cluster is re-initialized.