Metadata queries are not sent to same zone
Metadata queries for system stores are sent to lowest number node
in the cluster instead of the zone. Added a hack to the local pref
strategy if the client zone is set, use the zone local routing.
The code is very complicated (unnecesarily) did not clean it up as I
dont want to run it for all the scenarios and wanted to make a safe fix.