RO Hdfs fetcher allocates too much memory
1) Hdfs Fetcher in 1.0.4 uses ByteRangeInputStream. This class does not
override the method read(byte[], int , int). So it defaults to this
method from InputStream, which reads a character at a time from the
input stream. HttpInputStream for this method creates byte arrays for
each read.

So if you are download 2 TB data, the server will allocate/free 2 TB
data before the data is downloaded. This creates too much
garbage and new gen gets full in few milliseconds and GC happens. Though
GC are fast, this too much GC causes the latency to spike and causes JVM
to run out of Memory.

2) http://svn.apache.org/viewvc?view=revision&revision=1330500 fixed
this issue on April 2012 rather knowingly/unknowingly.

I tried upgrading to Hadoop latest but it brings in ProtoBuf 2.5.0 and
Avro 1.7. When I disabled the dependencies it failed at runtime
expecting protobuf 2.5.0 . I enabled only protobuf and it has no runtime
dependency on Avro 1.7. But I am saving that fix for a later day.

The branch is hadoop_Version_Upgrade which uses Hadoop 2.6.0 and
ProtoBuf 2.6.1