ConnectionException is not catastrophic
1) If a connection timesout or fails during protocol negotiation,
they are treated as normal errors instead of catastrophic errors.
Connection timeout was a regression from NIO connect fix. Protocol
negotiation timeout is a new change to detect the failed servers
faster.
2) When a node is marked down, the outstanding queued requests are
not failed and let them go through the connection creation cycle.
When there is no outstanding requests they can wait infinitely until
the next request comes up.
3) UnreachableStoreException is sometimes double wrapped. This causes
the catastrophic errors to be not detected accurately. Created an utility
method, when you are not sure if the thrown exception could be
UnreachableStoreException use this method, which handles this case
correctly.
4) In non-blocking connect if the DNS does not resolve the Java throws
UnresolvedAddressException instead of UnknownHostException. Probably an
issue in java. Also UnresolvedAddressException is not derived from IOException
but from IllegalArgumentException which is weird. Fixed the code to handle
this.
5) Tuned the remembered exceptions timeout to twice the connection timeout.
Previously it was hardcoded to 3 seconds, which was too aggressive when the
connection for some use cases where set to more than 5 seconds.

Added unit tests to verify all the above cases.