Report mmap/munmap errors using IOException
which the signature claims can be thrown but never was.
Ensure that the only callchain (MappedFileReader -> MemLock
-> mman) will actually handle and log IOException correctly.

Also, stop claiming that mlock() and munlock() throw IOException.
They never did and it would not be helpful, in fact we want to
ignore all errors from those functions because their normal
behavior in unprivileged processes is to fail.