Revert "Clean Set Quota Fix"
This reverts commit 3ff1cbe47a2206834c2cd6b77878dcccb3f0be67.

The unit tests are broken. On a deeper look, file backed
caching storage engine, does not have a version per key,
but at a file level. Any updates to the file with lower version
will be rejected.

So we have to go with super clocks for now. I am reverting the commit
to unblock the release of a new version.