Fork lift corrupts the data on schema mismatch
1) If the source and destination schema does not match, currently
forklift corrupts the destination data by streaming in bytes from
the source data.

After this commit, the forklift will fail when a schema mismatch
is detected. The old behavior if required can be achieved by
undocumented parameter ignore-schema-mismatch

Default of fork lift which forklifts all stores is changed to
fail when store name is not specified. I can't imagine a situation
where you want to forklift form one cluster to other often. If an
admin forgets to specify this parameter, they are forklifting the
entire cluster which is not definitely intended default.

Added 5 unit tests ( 3 for key mismatch and 2 for value mismatch).

Added pretty print functions to Compression and SerializerDefintion.