Client do not remove dead connections
When Server grace fully closes a connection, client ignores the
connection close. After this an operation that gets this connection
fails with EOFException as read/write returns -1 which causes the
operation to fail. Mostly this is fine as the backup operation generaly
succeeds, but when a cluster is bounced due to upgrade or other reasons,
and if a client ends up caching all dead connections, it causes the
operation to fail.

After a client operation is done, previously the socket is registered
with the selector for no-intention. For the next operation we start with
write.

With this change, the socket will be left in read mode. So FIN packet
from the server will be processed and the socket will be closed. Any
connection checkout from the pool needs to validate the connection.
There is a race condition as the returned connection from the cache
could still be killed later, that is handled as well.

Stream processors must be closed after the socket is closed.
Client will not be able to handle these close messages correctly.

Added more comments to the Client Request Executor.

Tests for dead connection clearing