CopyStats Regression
MessageDigest.digest() is read once, after it is read
the contents of the digest are reset.

Without knowing this, now started outputting checksum to stats file
which broke the normal flow.

Passing Around MessageDigest is dangerous because of this, so modified
the method to return byte arrays which are then passed around.