Update metadata version cluster, stores
At times metadata version on a cluster.xml and stores.xml drifts
The next metadata update instead of consolidating these versions
updates at few places and ignores at few other places.

This causes the client to not re-bootstrap correctly when the
cluster.xml or stores.xml is changed.

Now when a cluster.xml is changed, the version is synchronized
across the cluster to let the clients auto rebootstrap.

This fix merges the VectorClock on all the nodes to be updated
so that the Stores version will be updated correctly.

The old methods which does not take nodes as parameters are removed
and the public method exposes the nodes as parameters.