Fix RO JMX register/Unregister
1) Stores JMX are registered by JmxService class.
2) Previously ReadOnlyStorageConfiguration registered the same metrics
by prepending the NodeId. As Part of the commit

https://github.com/voldemort/voldemort/commit/2aec46a1edfbb6ab1eac4529d59ceb13f51e9fec#diff-005b79e324515c9e1045a61d0aee6d07

I fixed it and removed the NodeId. But I did not realize that this
caused the name collission and overwriting.

[18:50:45,445 voldemort.server.jmx.JmxService] WARN Overwriting mbean
voldemort.store.readonly:type=test2 [main]
[18:50:45,447 voldemort.server.jmx.JmxService] WARN Overwriting mbean
voldemort.store.readonly:type=test1 [main]

Now ReadOnlyStorageConfiguration does not register any JMX Metric at
all. Used the JConsole to verify that it is the same object registered
under two different names. Now the duplicate one is gone and the
warnings on the shutdown of Read Only server will be gone as well.