UnknownHostException is different between JDK 1.6 and JDK 1.8
http://docs.oracle.com/javase/7/docs/api/java/nio/channels/UnresolvedAddressException.html
has bad class hiearchy in JDK 1.6 seems like corrected in JDK 1.8

It is not instance of IO or Connect Exceptions.

This broke the test, now the test looks for one of the other exception.
The intention of the test is to keep the catastrophic errors up to date
so it is hacky, but OK.